+++ app/src/admin/Admin.css
@@ -1,0 +1,3 @@
+.admin-tabs {
+    margin-bottom: 20px;
+}
